refactor: refactored core/connections module
